move base gpio class to common
